Move recenter button back.
